Use null-coalesce instead of null-coalesce-equals
The use of the PHP 8.0 only null-coalesce-equals (`??=`) means that this file causes Psalm to see it as a parse error for projects not yet migrated to PHP 8. There is currently no way to avoid this, as psalm considers a parse error an unignorable error, even though in this case the file would never be included unless it actually is running on PHP 8.0.

The only way to avoid this is to set the entire project as PHP 8.0 but then you wouldn't be notified of legitimate issues when deploying to PHP 7.x.

It's a simple change from `$a ??= $b` to `$a = $a ?? $b` form.